Fix multiple definition link errors
The IRremote library uses an interesting technique to allow users to
configure the library's behavior, such as which timer to use. When you
include <IRremote.hpp>, besides the typical declarations, the actual
definitions or implementations of functions are included. This is great
because it means we can configure internal details of the library without
needing to rebuild the library or use a proper build system.

The catch? If you include <IRremote.hpp> in several translation units,
each translation unit gets a copy of the library's implementation. The
linker will notice this and emit multiple definition errors.

The solution? In most translation units, only include the declarations
by defining the 'USE_IRREMOTE_HPP_AS_PLAIN_INCLUDE' macro. Only in
'irremote.cpp' do we include the actual implementations.