(Try to) improve the rover's logic
The rover now looks in three directions before deciding what to do. It
turns slightly left if it's too close to a right wall and slightly right
if it's too close to a left wall. If it's facing a wall, it turns 90
degrees left or right before continuing.

In an attempt to improve the rover, I also implemented an exponential
average. The estimate for the current distance in each direction is a
weighted average of the previous estimate and the newest measurement.

Maybe in the future we'll implement Kalman filters ;) ;)