Measure distance in more directions
Instead of measuring in three directions -- left, center, and right --
measure the distance in five directions: left, mid-left, center,
mid-right, and right. Mid-left is 45 degrees between left and center,
and mid-right is 45 degrees between right and center.

We hope this improves the rover's behavior when it's not perfectly
aligned with the course, e.g., the rover's forward direction makes a 45
degree angle with the direction of the walls.